Fix Route Performance Issue in Wishlist and cart page
