+++ components/Loader.tsx
@@ -1,0 +1,11 @@
+
+import { TbLoader3 } from "react-icons/tb";
+
+export const Loader = () => {
+  return (
+    <div className="flex items-center justify-center h-[100px]">
+      <TbLoader3 className="animate-spin" size={32} />
+    </ div>
+  )
+
+}
